main cleaned up, seperated into providers, routes, navigation files.
